fixed: compile warnings
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@30767 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/GUIDialogAddonSettings.cpp
@@ -115,7 +115,7 @@ bool CGUIDialogAddonSettings::OnMessage(CGUIMessage& message)
       CGUIDialogBoxBase::OnMessage(message);
       int focusedControl = GetFocusedControlID();
       if (focusedControl >= CONTROL_START_SECTION && focusedControl < (int)(CONTROL_START_SECTION + m_totalSections) &&
-          focusedControl - CONTROL_START_SECTION != m_currentSection)
+          focusedControl - CONTROL_START_SECTION != (int)m_currentSection)
       { // changing section
         UpdateFromControls();
         m_currentSection = focusedControl - CONTROL_START_SECTION;
